Extract feature contract snapshot logic into core crate
