+++ crates/bitnet-feature-contract-core/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "bitnet-feature-contract-core"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "SRP core snapshot + drift contract types for policy/runtime feature alignment"
+include = ["src/**", "Cargo.toml"]
+
+[dependencies]
+
+[lints]
+workspace = true
