+++ crates/bitnet-feature-contract/Cargo.toml
@@ -13,7 +13,7 @@ description = "Compact feature and BDD profile contracts for BitNet tooling"
 include = ["src/**", "Cargo.toml"]
 
 [dependencies]
-bitnet-bdd-grid = { path = "../bitnet-bdd-grid", version = "0.2.1-dev" }
+bitnet-feature-contract-core = { path = "../bitnet-feature-contract-core", version = "0.2.1-dev" }
 bitnet-feature-matrix = { path = "../bitnet-feature-matrix", version = "0.2.1-dev" }
 
 [dev-dependencies]
